feat: add auth context with caching
